moving test_url to v21 dir and adding better compare
